backend/requirements.txt
django==3.2.16                       # Джанго
djangorestframework==3.14.0          # Джанго REST
psycopg2-binary==2.9.5               # Драйвер для работы с postgres
python-dotenv==0.21.0                # Для работы с env
Pillow==9.3.0                        # Для работы с изображениями
djoser==2.1.0                        # Для авторизации по токенам
django-filter==22.1                  # Для фильтрации набора запросов
